rename : TypeOption -> VoucherType
